requirements.txt
PySide6>=6.6.0
psutil>=5.9.0
Pillow>=10.0.0
